+++ R/estimate_carbon.R
@@ -122,6 +122,7 @@ estimate_carbon <- function(data, carbon_dir = here::here("carbon_code")) {
       PLT_CN,
       CONDID,
       STATUSCD,
+      RECONCILECD, #might be done with this column?
       DECAYCD,
       STANDING_DEAD_CD,
       SPCD,